@@ -2,25 +2,23 @@ var TagSwitcher = (function () {
   var $tagSwitcher, pollIntervalSeconds = 30, timeoutFunction;
 
   return {
-    init : function () {
+    init: function () {
       $tagSwitcher = $('select#tag');
 
-      if($tagSwitcher) {
+      if ($tagSwitcher) {
         $tagSwitcher.val($('#current_tag').val());
         $tagSwitcher.change(TagSwitcher.switchTags);
       }
     },
 
     switchTags: function () {
-      tag = $(this).val();
-      if(tag === '')
-        TagSwitcher.doRedirect('');
-      else
-        TagSwitcher.doRedirect("?tags=" + tag);
+      var tag = $(this).val();
+      var destination = tag === '' ? '' : "?tags=" + tag;
+      TagSwitcher.doRedirect(destination);
     },
 
     doRedirect: function (href) {
-      new_location = window.location.protocol + "//" + window.location.host + window.location.pathname + href;
+      var new_location = window.location.protocol + "//" + window.location.host + window.location.pathname + href;
       window.location.href = new_location;
     }
   };
