@@ -1,5 +1,6 @@
 class AddIterationStoryStateCountsToProjects < ActiveRecord::Migration
   def change
-    add_column :projects, :iteration_story_state_counts, :text, :default => "{}"
+    options = ActiveRecord::Base.connection.adapter_name.downcase.starts_with?("mysql") ? {} : { :default => "{}" }
+    add_column :projects, :iteration_story_state_counts, :text, options
   end
 end
