@@ -20,8 +20,12 @@ class TeamCityJsonPayload < Payload
   end
 
   def parse_url(content)
-    self.parsed_url = "http://#{remote_addr}:8111/viewType.html?buildTypeId=#{parse_build_type_id(content)}"
-    "http://#{remote_addr}:8111/viewLog.html?buildId=#{parse_build_id(content)}&tab=buildResultsDiv&buildTypeId=#{parse_build_type_id(content)}"
+    if content['buildStatusUrl']
+      self.parsed_url = content['buildStatusUrl']
+    else
+      self.parsed_url = "http://#{remote_addr}:8111/viewType.html?buildTypeId=#{parse_build_type_id(content)}"
+      "http://#{remote_addr}:8111/viewLog.html?buildId=#{parse_build_id(content)}&tab=buildResultsDiv&buildTypeId=#{parse_build_type_id(content)}"
+    end
   end
 
   def parse_build_type_id(content)
