@@ -19,7 +19,9 @@ class ProjectWorkloadHandler
   end
 
   def workload_failed(workload, e)
-    project.payload_log_entries.build(error_type: "#{e.class}", error_text: "#{e.message}", update_method: "Polling", status: "failed", backtrace: "#{e.message}\n#{e.backtrace.join("\n")}")
+    error_text = e.try(:message)
+    error_backtrace = e.try(:backtrace).try(:join,"\n")
+    project.payload_log_entries.build(error_type: "#{e.class}", error_text: "#{e.try(:message)}", update_method: "Polling", status: "failed", backtrace: "#{error_text}\n#{error_backtrace}")
     project.building = project.online = false
     project.save!
   end
@@ -32,8 +34,8 @@ private
     payload.status_content = status_content
     payload.build_status_content = build_status_content if project.build_status_url
 
-    payload_processor = PayloadProcessor.new(project, payload)
-    log = payload_processor.process
+    payload_processor = PayloadProcessor.new(project_status_updater: StatusUpdater.new)
+    log = payload_processor.process_payload(project: project, payload: payload)
     log.update_method = 'Polling'
     log.save!
 
@@ -45,7 +47,6 @@ private
     project.payload_log_entries.build(error_type: "#{e.class}", error_text: "#{e.message}", update_method: "Polling", status: "failed", backtrace: "#{e.message}\n#{e.backtrace.join("\n")}")
     project.online = false
     project.save!
-
   end
 
 end
