@@ -81,15 +81,20 @@ describe ProjectWorkloadHandler do
     let(:error) { double }
 
     before do
-      error.stub(:message).and_return("message")
       error.stub(:backtrace).and_return(["backtrace", "more"])
     end
 
     after { handler.workload_failed(workload, error) }
 
     it 'should add a log entry' do
+      error.stub(:message).and_return("message")
       project.payload_log_entries.should_receive(:build)
       .with(error_type: "RSpec::Mocks::Mock", error_text: "message", update_method: "Polling", status: "failed", :backtrace => "message\nbacktrace\nmore")
+    end
+
+    it 'should not call message on a failure when passed a String instead of an Exception' do
+      project.payload_log_entries.should_receive(:build)
+      .with(error_type: "RSpec::Mocks::Mock", error_text: "", update_method: "Polling", status: "failed", :backtrace => "\nbacktrace\nmore")
     end
 
     it 'should set building to false' do
