@@ -3,11 +3,11 @@ require "spec_helper"
 describe "Project" do
   describe "/validate_build_info" do
     it "returns log entry" do
-      project = FactoryGirl.create(:project)
-      ProjectUpdater.stub(:update).and_return(PayloadLogEntry.new(error_text: "Build unsuccessful"))
+      project = FactoryGirl.create(:project, jenkins_build_name: 'twitter-for-dogs')
+      fake_jenkins!
 
-      post "/projects/validate_build_info", project: project.attributes.merge(auth_password: "password")
-      expect(response.body).to match /Build unsuccessful/
+      patch "/projects/validate_build_info", project: project.attributes.merge(auth_password: "password")
+      expect(response.body).to match /Error parsing content for build name twitter-for-dogs/
     end
   end
 
@@ -18,5 +18,10 @@ describe "Project" do
       post "/projects/validate_tracker_project", { id: project.id, auth_token: 'token', project_id: project.id }
       expect(response.status).to be(202)
     end
+  end
+
+  def fake_jenkins!
+    FakeWeb.register_uri(:get, "http://www.example.com/job/twitter-for-dogs/rssAll", body: File.new('spec/fixtures/jenkins_atom_examples/invalid_xml.atom'))
+    FakeWeb.register_uri(:get, "http://www.example.com/cc.xml", body: File.new('spec/fixtures/jenkins_atom_examples/invalid_xml.atom'))
   end
 end
